+++ SourceCode/Data/LayoutModule.cs
@@ -36,6 +36,13 @@ public static class LayoutModuleExtensions
     public static bool HasLayoutStationId(this LayoutModule it) => it.LayoutStationId.HasValue;
     public static bool HasLayoutStation(this LayoutModule it) => it.LayoutStation is not null;
     public static bool HasCargoCustomers(this LayoutModule it) => it.LayoutStation?.Station is not null && it.LayoutStation.Station.HasCargoCustomers;
+    public static double TotalLength(this LayoutModule it) => it.Module?.TotalLength() ?? 0.0;
+    public static double TotalLength(this IEnumerable<LayoutModule> modules)
+    {
+        double sum = 0.0;
+        foreach (var module in modules) sum += module.TotalLength();
+        return sum;
+    }
 }